+++ Loader/src/ModEntry.cs
@@ -663,7 +663,7 @@ namespace Sheepy.Modnix {
       public string[]  Mods;
       public string[]  Preloads;
       public DllMeta[] Dlls;
-      public Dictionary<string,object>[] Actions;
+      public IDictionary<string,object>[] Actions;
 
       public   string  ConfigType;
       internal string  DefaultConfigText;
@@ -789,7 +789,7 @@ namespace Sheepy.Modnix {
          if ( val.Length == 0 ) val = null;
       }
 
-      internal static void NormDictArray ( ref Dictionary<string,object>[] val ) {
+      internal static void NormDictArray ( ref IDictionary<string,object>[] val ) {
          if ( val == null ) return;
          for ( int i = val.Length - 1 ; i >= 0 ; i-- ) {
             var dict = val[i];